Generative Models/AI-Query-editor/requirements.txt
flask==2.3.3
google-generativeai==0.3.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pymysql==1.1.0
werkzeug==2.3.7
gunicorn==21.2.0
cryptography==41.0.7
protobuf==4.25.1
requests==2.31.0